// nextjs-frontend/app/openapi-client/types.gen.ts
export type NotebookPublic = {
  title: string;
  description?: string | null;
  owner_id: string;
  id: string;
  created_at: string;
  updated_at: string;
  /**
   * Number of sources linked to this notebook
   */
  source_count?: number;
};